% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.cgps_cw}
\alias{autoplot.cgps_cw}
\title{A helper function for cgps_cw object}
\usage{
\method{autoplot}{cgps_cw}(object, ...)
}
\arguments{
\item{object}{A cgps_cw object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot cgps_cw object using ggplot2 package.
}
\keyword{internal}
