% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.cgps_erf}
\alias{autoplot.cgps_erf}
\title{A helper function for cgps_erf object}
\usage{
\method{autoplot}{cgps_erf}(object, ...)
}
\arguments{
\item{object}{A cgps_erf object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot cgps_erf object using ggplot2 package.
}
\keyword{internal}
