% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.cgps_gps}
\alias{autoplot.cgps_gps}
\title{A helper function for cgps_gps object}
\usage{
\method{autoplot}{cgps_gps}(object, ...)
}
\arguments{
\item{object}{A cgps_gps object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot cgps_gps object using ggplot2 package.
}
\keyword{internal}
