% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.cgps_pspop}
\alias{autoplot.cgps_pspop}
\title{A helper function for cgps_pspop object}
\usage{
\method{autoplot}{cgps_pspop}(object, ...)
}
\arguments{
\item{object}{A cgps_pspop object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot cgps_pspop object using ggplot2 package.
}
\keyword{internal}
