% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matching.R
\name{create_matching}
\alias{create_matching}
\title{Create pseudo population using matching casual inference approach}
\usage{
create_matching(
  .data,
  exposure_col_name,
  matching_fn,
  dist_measure = dist_measure,
  gps_density = gps_density,
  delta_n = delta_n,
  scale = scale,
  bin_seq = NULL,
  nthread = 1
)
}
\arguments{
\item{.data}{TBD}

\item{gps_density}{Model type which is used for estimating GPS value, including
\code{normal} (default) and \code{kernel}.}

\item{bin_seq}{Sequence of w (treatment) to generate pseudo population. If
NULL is passed the default value will be used, which is
\code{seq(min(w)+delta_n/2,max(w), by=delta_n)}.}

\item{nthread}{Number of available cores.}
}
\value{
Returns data.table of matched set.
}
\description{
Generates pseudo population based on matching casual inference method.
}
\keyword{internal}
