% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_semipmetric_erf.R
\name{estimate_semipmetric_erf}
\alias{estimate_semipmetric_erf}
\title{Estimate semi-exposure-response function (semi-ERF).}
\usage{
estimate_semipmetric_erf(formula, family, data, ...)
}
\arguments{
\item{formula}{a vector of outcome variable in matched set.}

\item{family}{a description of the error distribution (see ?gam).}

\item{data}{dataset that formula is build upon Note that there should be a
\code{counter_weight} column in this data.).}

\item{...}{Additional parameters for further fine tuning the gam model.}
}
\value{
returns an object of class gam
}
\description{
Estimates the smoothed exposure-response function using a generalized
additive model with splines.
}
\details{
This approach uses Generalized Additive Model (gam) using mgcv package.
}
\keyword{internal}
