% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cgps_cw}
\alias{plot.cgps_cw}
\title{Extend generic plot functions for cgps_cw class}
\usage{
\method{plot}{cgps_cw}(x, ...)
}
\arguments{
\item{x}{A cgps_cw object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot2 object, invisibly. This function is called for side effects.
}
\description{
A wrapper function to extend generic plot functions for cgps_cw class.
}
\details{
Additional parameters:
\itemize{
\item \emph{every_n}: Puts label to ID at every n interval (default = 10)
\item \emph{subset_id}: A vector of range of ids to be included in the plot
(default = NULL)
}
}
