% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cgps_gps}
\alias{plot.cgps_gps}
\title{Extend generic plot functions for cgps_gps class}
\usage{
\method{plot}{cgps_gps}(x, ...)
}
\arguments{
\item{x}{A cgps_gps object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot2 object, invisibly. This function is called for side effects.
}
\description{
A wrapper function to extend generic plot functions for cgps_gps class.
}
