% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cgps_pspop}
\alias{plot.cgps_pspop}
\title{Extend generic plot functions for cgps_pspop class}
\usage{
\method{plot}{cgps_pspop}(x, ...)
}
\arguments{
\item{x}{A cgps_pspop object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot2 object, invisibly. This function is called for side effects.
}
\description{
A wrapper function to extend generic plot functions for cgps_pspop class.
}
\details{
\subsection{Additional parameters}{
\itemize{
\item \emph{include_details}: If set to TRUE, the plot will include run details (Default = FALSE).
}
}
}
