% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_counter_weight.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data}
\usage{
preprocess_data(data, trim_quantiles, exposure_col)
}
\arguments{
\item{data}{description}

\item{exposure_col}{Column name that is used for exposure.}

\item{trim_quntiles}{description}
}
\value{
A list with preprocessed and original data.
}
\description{
Preprocess data to isolate extra details
}
\keyword{internal}
