% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cgps_cw}
\alias{print.cgps_cw}
\title{Extend print function for cgps_cw object}
\usage{
\method{print}{cgps_cw}(x, ...)
}
\arguments{
\item{x}{A cgps_cw object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Extend print function for cgps_cw object
}
