% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cgps_erf}
\alias{print.cgps_erf}
\title{Extend print function for cgps_erf object}
\usage{
\method{print}{cgps_erf}(x, ...)
}
\arguments{
\item{x}{A cgps_erf object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Extend print function for cgps_erf object
}
