% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cgps_gps}
\alias{print.cgps_gps}
\title{Extend print function for cgps_gps object}
\usage{
\method{print}{cgps_gps}(x, ...)
}
\arguments{
\item{x}{A cgps_gps object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Extend print function for cgps_gps object
}
