% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cgps_cw}
\alias{summary.cgps_cw}
\title{print summary of cgps_cw object}
\usage{
\method{summary}{cgps_cw}(object, ...)
}
\arguments{
\item{object}{A cgps_cw object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data
}
\description{
print summary of cgps_cw object
}
