% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cgps_erf}
\alias{summary.cgps_erf}
\title{print summary of cgps_erf object}
\usage{
\method{summary}{cgps_erf}(object, ...)
}
\arguments{
\item{object}{A cgps_erf object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data
}
\description{
print summary of cgps_erf object
}
