% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cgps_gps}
\alias{summary.cgps_gps}
\title{print summary of cgps_gps object}
\usage{
\method{summary}{cgps_gps}(object, ...)
}
\arguments{
\item{object}{A cgps_gps object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data
}
\description{
print summary of cgps_gps object
}
