% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cgps_pspop}
\alias{summary.cgps_pspop}
\title{print summary of cgps_pspop object}
\usage{
\method{summary}{cgps_pspop}(object, ...)
}
\arguments{
\item{object}{A cgps_pspop object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data
}
\description{
print summary of cgps_pspop object
}
