% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_erf.R
\name{estimate_erf}
\alias{estimate_erf}
\title{Estimate Exposure Response Function}
\usage{
estimate_erf(.data, .formula, weights_col_name, model_type, w_vals, ...)
}
\arguments{
\item{.data}{TBD}

\item{.formula}{TBD}

\item{weights_col_name}{TBD}

\item{model_type}{tBD}

\item{...}{TBD}
}
\value{
TBD
}
\description{
A short description...
}
\examples{
TBD
}
